kernel-secmon/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-multipart
websockets
google-generativeai
python-dotenv
